requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
python-multipart
sqlalchemy
asyncmy
python-dotenv
email-validator
bcrypt>=4.0.0